requirements.txt
biopython==1.79
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
# -f https://data.dgl.ai/wheels/repo.html
# dgl-cu111==0.7.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.9.0+cu111
# -f https://download.pytorch.org/whl/torch_stable.html
# torchaudio==0.9.0
# -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.10.0+cu111
typing-extensions==3.10.0.0
urllib3==1.26.5
wandb==0.12.1
tqdm==4.62.2

# rust-dbg
lz4==4.0.0

#
hydra-core==1.2.0
omegaconf==2.2.2
typeguard===2.12.1
